refactor(budget_disclosure): graduate to top-level layout, add docs-site page
Merge `main` and relocate the capability from `experimental/budget_disclosure`
to the canonical top-level `pydantic_ai_harness/budget_disclosure/`, matching the
layout from #347/#354. Drops the experimental import warning, fixes import paths,
and moves tests. Adds the `docs/budget-disclosure.md` page (registered in
nav.json, index.md, and the parity gate) and a README matrix row. No behavior change.
